test(datetimepicker): add tests structure and test value prop
